Standardize layout of item/category listings.
Added partials and CSS to create a general list view of items and categories. They are similar in that they each have a name and image.
They are dis-similar in that items include the Add To Cart button. This is why I use a nested partial solution for items.
This change involved adding the image_url column to the categories table. I also modified the seed data for both categories and items.