Fix item_update_bug
Fixed local variable name in _form which was :html_method but should be just 'html_method'. Also corrected the path set in the form to be admin_item_path(@item) instead of admin_items_path... and fixed the admin/items controller to find the item by_dao instead of the item id which wasn't actually provided in the params.