Fix bug stopping new items to be created
I think I caused this bugged when fixing the update item bug from before. The create and update actions require different action paths, but they use the same form in which we define that path, so if we defined it to work for create, the update didn't work, and if we defined it properly for update, create didn't work. The solution is a location 'path' variable sent in when the _form partial is called. So, I edited both files adding in the path and changed the _form partial to accept that local variable. Now they both work and the tests are all passing.